requirements.txt
annotated-types==0.7.0
anyio==4.12.1
blinker==1.9.0
cachetools==6.2.6
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.1.8
colorama==0.4.6
cryptography==47.0.0
deprecation==2.1.0
exceptiongroup==1.3.1
Flask==3.1.3
flask-cors==6.0.2
fsspec==2025.10.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.13
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.0
multidict==6.7.1
packaging==26.2
postgrest==2.29.0
propcache==0.4.1
pycparser==2.23
pydantic==2.13.3
pydantic_core==2.46.3
Pygments==2.20.0
pyiceberg==0.10.0
PyJWT==2.12.1
pyparsing==3.3.2
pyroaring==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
realtime==2.29.0
requests==2.32.5
rich==14.3.4
six==1.17.0
sortedcontainers==2.4.0
storage3==2.29.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.29.0
supabase-auth==2.29.0
supabase-functions==2.29.0
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==15.0.1
Werkzeug==3.1.8
yarl==1.22.0
zipp==3.23.1
